+++ BE/src/test/java/team05/codesquad/issuetracker/repository/IssueRepositoryTest.java
@@ -1,11 +1,10 @@
 package team05.codesquad.issuetracker.repository;
 
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import team05.codesquad.issuetracker.domain.Issue;
+import team05.codesquad.issuetracker.domain.issue.Issue;
 
 import static org.assertj.core.api.Assertions.*;
